Create spatial driver data for model input

Regularizes spatial input data to a common equal area grid. It also performs sanity checks on the inputs to check if observations cover the spatial grid. Data can be provided as arguments, or if no inputs are provided the data will be downloaded in the background.

hr_drivers(
  track,
  crs = "ESRI:54009",
  buffer = 7,
  window = 11,
  path,
  na_value = NA,
  overwrite = TRUE
)

Arguments

track

observed locations track

crs

an equal area CRS projection reference (default = ESRI:54009, or the Mollweide global equal area approximation)

buffer

buffer (in km, default = 11) around the bio-logging track file to accommodate for a sufficiently large home range area. Value should be adjusted to the target species.

window

window size for resampled forest cover values, number of pixels of the window (uneven value, default = 11 or approx. 300m)

path

path and filename where to save the raster map data. When using a temporary file use file.path(tempdir(), "raster_drivers.tif") or similar

na_value

value with which to substitute NA values (e.g. -9999)

overwrite

overwrite output if the path name is the same (default = TRUE)

Value

nested list with the regularized data and observed track, also saves the used projection crs

Details

It is encouraged to save the data to file, rather than use a returned object. This limits the memory footprint of operations.